% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_pdf.R
\name{beam_this_rmd}
\alias{beam_this_rmd}
\title{beam_this_rmd()}
\usage{
beam_this_rmd(themecolor = "sapphire", fontcolor = "graphite",
  toc = FALSE, incremental = FALSE, fig_width = 9, fig_height = 6,
  fig_crop = TRUE, fig_caption = TRUE, keep_tex = TRUE,
  pandoc_args = NULL, highlight = "haddock",
  latex_engine = "xelatex")
}
\arguments{
\item{themecolor}{Sets presentation theme color (available: white, black, sapphire, graphite)}

\item{fontcolor}{Sets presentation font color (available: white, black, sapphire, graphite, mokka)}

\item{toc}{Logical indicating whether to create 'Table of Contents'}

\item{incremental}{Logical indicating whether to render bullet points incrementally. Defaults to FALSE. To reverse-order, precede item with `>'}

\item{fig_width}{Default figure width, in inches}

\item{fig_height}{Default figure height, in inches}

\item{fig_crop}{Logical whether to crop figures}

\item{fig_caption}{Logical indicating whether to render figure captions}

\item{keep_tex}{Logical indicating whether to keep .tex version of file (Recommended not switching to FALSE)}

\item{pandoc_args}{Optional, additional args to pass on to pandoc}

\item{highlight}{Highlight style, if "default" then "haddock" is used}

\item{latex_engine}{Indicating what LaTeX engine to use}
}
\description{
Converting .Rmd file into beautiful MiNI WUT themed beamer presentation
}
\details{
Only specified pairs of themecolor and fontcolor attributes are available (see: github/mckraqs/miniBeamer)
}
