% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_business_card.R
\name{rmd_to_business_card}
\alias{rmd_to_business_card}
\title{Create business cards}
\usage{
rmd_to_business_card(
  template = pkg_resource(main_dir =
    "rmarkdown/templates/rmd_to_business_card/resources", "html", "card.html")
)
}
\arguments{
\item{template}{The path to the Pandoc template to convert Markdown to HTML.}
}
\value{
An R Markdown output format.
}
\description{
Create business cards
}
