% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnw_to_rmd.R
\name{rnw_to_rmd}
\alias{rnw_to_rmd}
\title{rnw_to_rmd()}
\usage{
rnw_to_rmd(filename)
}
\arguments{
\item{filename}{String that indicates the location and name of the .Rnw file to be converted.}
}
\description{
Converting .Rnw file into .Rmd
}
\details{
In order for the conversion to work properly the working directory has to be in the folder containing the .Rnw file.
}
